refactor: split files exceeding size limits (#212)
Extract bump logic from standard-version/src/lib.rs into bump.rs,
scan/update functions from version_file.rs into scan.rs, and tag
queries from git-std/src/git/query.rs into tag.rs. All re-exports
preserved so public API is unchanged.
